requirements.txt
gradio
langchain_groq
langchain_core
langchain_community
chromadb
pypdf
sentence-transformers
python-dotenv
